backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib
httpx
google-auth
google-auth-oauthlib
google-api-python-client
openpyxl
python-dotenv
python-multipart
groq
psycopg2-binary
gunicorn
